+++ src/mytodo.domain/Handlers/User/DeleteUserRequestHandler.cs
@@ -1,8 +1,10 @@
 using MediatR;
 using mytodo.domain.Repository;
+using mytodo.shareable.Excecoes;
 using mytodo.shareable.Requests.User;
 using mytodo.shareable.Responses.User;
 using OperationResult;
+using static mytodo.shareable.Enums.Cadastro;
 
 namespace mytodo.domain.Handlers.User;
 
@@ -10,19 +12,31 @@ public class DeleteUserRequestHandler : IRequestHandler<DeleteUserRequest, Resul
 {
     private readonly IUserRepository _userRepository;
     private readonly IUnitOfWork _unitOfWork;
-    
+
     public DeleteUserRequestHandler(IUserRepository userRepository, IUnitOfWork unitOfWork)
     {
         _userRepository = userRepository;
         _unitOfWork = unitOfWork;
     }
-    
+
     public async Task<Result<DeleteUserResponse>> Handle(DeleteUserRequest request, CancellationToken cancellationToken)
     {
         var user = await _userRepository.GetUserByIdAsync(request.UserId);
 
-        await _userRepository.DeleteUserAsync(user);
-        await _unitOfWork.SaveChangesAsync(cancellationToken);
+        if (user == null)
+        {
+            return Result.Error<DeleteUserResponse>(new ExcecaoAplicacao(BuscaNaoEncontrada));
+        }
+
+        try
+        {
+            await _userRepository.DeleteUserAsync(user);
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+        }
+        catch
+        {
+            return Result.Error<DeleteUserResponse>(new ExcecaoAplicacao(FalhaAoDeletar));
+        }
 
         return Result.Success(new DeleteUserResponse(user.UserId, user.Username));
     }
